+++ NEWS.md
@@ -1,3 +1,10 @@
+CohortPrevalence v0.0.3
+=======================
+* add demographic constraint to population: limit by age range or specific gender
+* fix pd4 bug
+* fix prevalence results collect bug
+* add rassen incidence calculator
+
 CohortPrevalence v0.0.2
 =======================
 * add ability to specify a multi year span as an analysis: ie 2020-2024